requirements.txt
async-generator==1.10
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.1
h11==0.13.0
idna==3.3
lxml==4.8.0
outcome==1.1.0
progress==1.6
progress-bar==8
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
python-docx==0.8.11
random-user-agent==1.0.1
requests==2.27.1
selenium==4.1.3
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.20.0
trio-websocket==0.9.2
typing-extensions==4.2.0
undetected-chromedriver==3.1.5.post4
urllib3==1.26.9
websockets==10.3
wsproto==1.1.0